backend/requirements.txt
fasttext-wheel
flask
flask-cors
torch
transformers
huggingface-hub